refactor(listener): extract memory commands
